+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- add shadow-(un)ban, deactivate
+- add shadow-(un)ban, deactivate, registration token
 
 ## [v0.1.1] - 2022-08-26